+++ packages/cortex-react/src/components/ProgressBar/ProgressBar.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Progress } from './Progress';
+
+export interface ProgressBarProps {
+  value?: number;
+  intent?: 'default' | 'success' | 'warning' | 'info' | 'error';
+  infinite?: boolean;
+}
+
+export const ProgressBar = ({
+  value = 50,
+  intent = 'default',
+  infinite = false,
+}: ProgressBarProps) => {
+  return <Progress value={value} intent={intent} infinite={infinite} />;
+};